pyproject.toml
[project]
name = "desfire"
version = "0.11"
description = "DESFire library for python"
authors = [
    {name = "Daniel Herrmann",email = "daniel.herrmann@makerspace-darmstadt.de"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "enum34 (>=1.1.10,<2.0.0)",
    "pyscard (>=2.2.1,<3.0.0)",
    "pydes (>=2.0.1,<3.0.0)",
    "pycryptodome (>=3.21.0,<4.0.0)"
]

[tool.poetry]
packages = [
    { include = "desfire", from = "src" }
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
mypy = "^1.14.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true
src = ["src"]

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
ignore = [
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E402", # module level import not at top of file
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP006", # use-pep585-annotation
    "UP007", # use-pep604-annotation
    "E741",  # Ambiguous variable name
    # "UP035", # deprecated-assertion
    "PLE1205"  # PLE1205 due to custom logger implementation
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401", # unused import
    "F403", # star imports
]

[tool.ruff.lint.isort]
case-sensitive = true
detect-same-package = true
known-first-party = ["backend", "worker", "helper"]

[tool.mypy]
disable_error_code = "import-untyped"
enable_incomplete_feature = ["NewGenericSyntax"]

